+++ src/Nono.Engine/Extensions/CuesExtensions.cs
@@ -4,10 +4,9 @@ namespace Nono.Engine.Extensions
 {
     public static class CuesExtensions
     {
-        public static ushort Sum(this ReadOnlySpan<ushort> span)
+        public static int Sum(this ReadOnlySpan<int> span)
         {
-            // Sum of all cues should not be bigger then a max row count
-            ushort sum = 0;
+            int sum = 0;
             for (int i = 0; i < span.Length; i++)
             {
                 sum += span[i];
@@ -16,9 +15,9 @@ namespace Nono.Engine.Extensions
             return sum;
         }
 
-        public static ushort Max(this ReadOnlySpan<ushort> span)
+        public static int Max(this ReadOnlySpan<int> span)
         {
-            ushort max = span[0];
+            int max = span[0];
             for (int i = 1; i < span.Length; i++)
             {
                 var element = span[i];
